Move setting stage holder to the core folder
