+++ web/src/app/content-management/schedules/schedules.module.ts
@@ -4,7 +4,7 @@ import {SingleScheduleComponent} from "./single-schedule/single-schedule.compone
 import {BrowserModule} from "@angular/platform-browser";
 import {AutocompleteModule} from "../../shared/autocomplete/autocomplete.module";
 import {SchedulesService} from "./schedules.service";
-import {SettingStateHolderService} from "../../settings/setting-state-manager/settings-state-holder.service";
+import {SettingStateHolderService} from "../../core/setting-state-manager/settings-state-holder.service";
 import {SchedulesComponent} from "./schedules.component";
 import {CheckboxMultiselectorModule} from "../../shared/checkbox-multiselector/checkbox-multiselector.module";
 import { DaysOfWeekMultiselectorComponent } from './single-schedule/days-of-week-multiselector/days-of-week-multiselector.component';
