+++ web/src/app/content-management/schedules/single-schedule/single-schedule.component.ts
@@ -2,7 +2,7 @@ import {Component, OnInit, Input, Output, EventEmitter} from '@angular/core';
 import {SchedulesService} from "../schedules.service";
 import {Schedule} from "../entities/schedule";
 import {EventTime} from "../entities/event-time";
-import {SettingStateHolderService} from "../../../settings/setting-state-manager/settings-state-holder.service";
+import {SettingStateHolderService} from "../../../core/setting-state-manager/settings-state-holder.service";
 import {Setting} from "../../../settings/entities/setting";
 import {ValidationResult} from "../entities/validation-result";
 import {Periodicity} from '../../../core/enums/periodicity';
