+++ web/src/app/settings/setting-creator/setting-creator.component.ts
@@ -2,7 +2,7 @@ import {Component, Output, EventEmitter, Input, OnInit} from '@angular/core';
 import {SettingsService} from "../settings.service";
 import {Setting} from "../entities/setting";
 import {NotificationService} from "../../shared/notifications/notification.service";
-import {SettingStateHolderService} from "../setting-state-manager/settings-state-holder.service";
+import {SettingStateHolderService} from "../../core/setting-state-manager/settings-state-holder.service";
 
 import * as _ from 'lodash';
 import {SchedulesService} from "../../content-management/schedules/schedules.service";
