+++ web/src/app/settings/settings-priority.helper.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import {SettingsService} from "./settings.service";
-import {SettingStateHolderService} from "./setting-state-manager/settings-state-holder.service";
+import {SettingStateHolderService} from "../core/setting-state-manager/settings-state-holder.service";
 import {Setting} from "./entities/setting";
 import {Schedule} from "../content-management/schedules/entities/schedule";
 import {Periodicity} from '../core/enums/periodicity';
